pyproject.toml
[project]
name = "code-quality-intelligence"
version = "0.1.0"
description = "An intelligent code quality analysis tool with interactive Q&A capabilities"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "pygments>=2.16.0",
    "tree-sitter>=0.20.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-core>=0.1.0",
    "langchain-ollama>=0.1.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "gitpython>=3.1.0",
    "radon>=6.0.0",
    "bandit>=1.7.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "redis>=5.0.0",
    "hiredis>=2.2.0",
]

[project.scripts]
codet = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
